Store the number of balls thrown
