+++ src/panel/states.ts
@@ -113,51 +113,6 @@ export function isStateAboveGround(state: States): boolean {
     );
 }
 
-export function ballCaught(pet: IPetType) {
-    vscode.window.showInformationMessage('Ball Caught');
-    const filePath = path.join(
-        path.dirname(__dirname),
-        'vscode-pets.stats.txt',
-    );
-    let ballCaughtCount = 0;
-
-    if (!fs.existsSync(filePath)) {
-        try {
-            const stats = { ballCaught: 0 };
-            const statsFileContent = JSON.stringify(stats, null, 2);
-            fs.writeFileSync(filePath, statsFileContent, {
-                encoding: 'utf-8',
-            });
-        } catch (e) {
-            vscode.window.showErrorMessage(
-                `Failed to create ${filePath}: ${e}`,
-            );
-        }
-    }
-
-    try {
-        const statsFileContent = fs.readFileSync(filePath, {
-            encoding: 'utf-8',
-        });
-        const stats = JSON.parse(statsFileContent);
-        ballCaughtCount = stats.ballCaught || 0;
-    } catch (e) {
-        vscode.window.showErrorMessage(`Failed to read ${filePath}: ${e}`);
-    }
-
-    ballCaughtCount++;
-
-    try {
-        const stats = { ballCaught: ballCaughtCount };
-        const statsFileContent = `Pet has caught ${stats.ballCaught} balls`;
-        fs.writeFileSync(filePath, statsFileContent, {
-            encoding: 'utf-8',
-        });
-    } catch (e) {
-        vscode.window.showErrorMessage(`Failed to update ${filePath}: ${e}`);
-    }
-}
-
 export function resolveState(state: string, pet: IPetType): IState {
     switch (state) {
         case States.sitIdle:
