+++ webpack.config.js
@@ -25,12 +25,6 @@ const desktopConfig = {
   ],
   resolve: {
     extensions: [".ts", ".tsx", ".js"],
-		fallback: {
-			fs: false,
-			vscode: false,
-			path: false,
-			path: require.resolve('path-browserify')
-		}
   },
   module: {
     rules: [{
